content/canvas/test/test_2d.imageData.get.sourcepos.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.imageData.get.sourcepos</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ctx.fillStyle = '#000';
ctx.fillRect(0, 0, 100, 50);
ctx.fillStyle = '#fff';
ctx.fillRect(20, 10, 60, 30);
var imgdata1 = ctx.getImageData(10, 5, 1, 1);
ok(imgdata1.data[0] < 50, "imgdata1.data[\""+(0)+"\"] < 50");
var imgdata2 = ctx.getImageData(30, 20, 1, 1);
ok(imgdata2.data[0] > 200, "imgdata2.data[\""+(0)+"\"] > 200");

if (!_deferred) SimpleTest.finish();
});
</script>
